ENH: allow multipart hdf5
