+++ tests/test_tiled_writer.py
@@ -315,7 +315,13 @@ def collect_plan(*objs, name="primary"):
 
 
 @pytest.mark.parametrize(
-    "fname", ["internal_events", "external_assets", "external_assets_legacy"]
+    "fname",
+    [
+        "internal_events",
+        "external_assets",
+        "external_assets_legacy",
+        "external_assets_multipart_hdf5",
+    ],
 )
 @pytest.mark.parametrize("batch_size", [0, 1, 1000, None])
 def test_with_correct_sample_runs(client, batch_size, external_assets_folder, fname):
